Déplace la vue bac dans commun.
