--- src/situations/controle/vues/bac.js
@@ -1,38 +1,0 @@
-import 'controle/styles/bac.scss';
-
-import { CHANGEMENT_ETAT_SURVOLE } from 'commun/modeles/bac';
-
-export default class VueBac {
-  constructor (bac) {
-    this.bac = bac;
-  }
-
-  affiche (pointInsertion, $) {
-    function creeElementBac (categorie, { x, y }, { largeur, hauteur }, { largeurParent, hauteurParent }) {
-      const $element = $('<div class="bac"></div>');
-      $element.css('left', x * largeurParent / 100);
-      $element.css('top', y * hauteurParent / 100);
-      $element.width(largeur * largeurParent / 100).height(hauteur * hauteurParent / 100);
-      return $element;
-    }
-
-    const $elementParent = $(pointInsertion);
-    const dimensionsElementParent = {
-      largeurParent: $elementParent.width(),
-      hauteurParent: $elementParent.height()
-    };
-    this.$bac = creeElementBac(
-      this.bac.categorie(),
-      this.bac.position(),
-      this.bac.dimensions(),
-      dimensionsElementParent
-    );
-
-    $elementParent.append(this.$bac);
-    this.bac.on(CHANGEMENT_ETAT_SURVOLE, (etat) => this.changeEtatSurvole(etat));
-  }
-
-  changeEtatSurvole (etatSurvole) {
-    this.$bac.toggleClass('survole', etatSurvole);
-  }
-}
